==> 00_00_r_env_init.Rmd <==
---
title: "00_00_r_env_init"
author: "Puvvula"
date: "2024-03-19"
output: pdf_document
---

```{r}
library(pacman)
p_load(tidyverse, 
       parallel, foreach, doParallel, 
       raster, sf)
```


#directories
```{r}
data_path <- "D:/pa_births/formatted_data/"
result_path <- "D:/pa_births/results/"
```
